+++ src/linalg/blas/axpy.zig
@@ -7,12 +7,12 @@ const ops = @import("../../ops.zig");
 const blas = @import("../blas.zig");
 
 pub fn axpy(
-    n: isize,
+    n: i32,
     alpha: anytype,
     x: anytype,
-    incx: isize,
+    incx: i32,
     y: anytype,
-    incy: isize,
+    incy: i32,
     ctx: anytype,
 ) !void {
     const Al: type = @TypeOf(alpha);
@@ -24,24 +24,24 @@ pub fn axpy(
 
     if (ops.eq(alpha, 0, .{}) catch unreachable) return;
 
-    var ix: isize = if (incx < 0) (-n + 1) * incx else 0;
-    var iy: isize = if (incy < 0) (-n + 1) * incy else 0;
+    var ix: i32 = if (incx < 0) (-n + 1) * incx else 0;
+    var iy: i32 = if (incy < 0) (-n + 1) * incy else 0;
     if (comptime types.isArbitraryPrecision(C)) {
         if (comptime types.isArbitraryPrecision(Y)) {
             // Orientative implementation for arbitrary precision types
             var temp: C = try ops.init(C, ctx);
             defer ops.deinit(temp, ctx);
-            for (0..scast(usize, n)) |_| {
+            for (0..scast(u32, n)) |_| {
                 try ops.mul_(
                     &temp,
                     alpha,
-                    x[scast(usize, ix)],
+                    x[scast(u32, ix)],
                     ctx,
                 );
 
                 try ops.add_(
-                    &y[scast(usize, iy)],
-                    y[scast(usize, iy)],
+                    &y[scast(u32, iy)],
+                    y[scast(u32, iy)],
                     temp,
                     ctx,
                 );
@@ -58,13 +58,13 @@ pub fn axpy(
         if (comptime types.isArbitraryPrecision(Y)) {
             @compileError("zml.linalg.blas.axpy not implemented for arbitrary precision types yet");
         } else {
-            for (0..scast(usize, n)) |_| {
+            for (0..scast(u32, n)) |_| {
                 ops.add_( // y[iy] += alpha * x[ix]
-                    &y[scast(usize, iy)],
-                    y[scast(usize, iy)],
+                    &y[scast(u32, iy)],
+                    y[scast(u32, iy)],
                     ops.mul(
                         alpha,
-                        x[scast(usize, ix)],
+                        x[scast(u32, ix)],
                         ctx,
                     ) catch unreachable,
                     ctx,
