+++ src/linalg/blas/dot_sub.zig
@@ -7,11 +7,11 @@ const ops = @import("../../ops.zig");
 const blas = @import("../blas.zig");
 
 pub fn dot_sub(
-    n: isize,
+    n: i32,
     x: anytype,
-    incx: isize,
+    incx: i32,
     y: anytype,
-    incy: isize,
+    incy: i32,
     ret: anytype,
     ctx: anytype,
 ) !void {
@@ -24,18 +24,18 @@ pub fn dot_sub(
 
     if (n <= 0) return blas.Error.InvalidArgument;
 
-    var ix: isize = if (incx < 0) (-n + 1) * incx else 0;
-    var iy: isize = if (incy < 0) (-n + 1) * incy else 0;
+    var ix: i32 = if (incx < 0) (-n + 1) * incx else 0;
+    var iy: i32 = if (incy < 0) (-n + 1) * incy else 0;
     if (comptime types.isArbitraryPrecision(R)) {
         if (comptime types.isArbitraryPrecision(C)) {
             // Orientative implementation for arbitrary precision types
             var temp: C = try ops.init(C, ctx);
             defer ops.deinit(temp, ctx);
-            for (0..scast(usize, n)) |_| {
+            for (0..scast(u32, n)) |_| {
                 try ops.mul_(
                     &temp,
-                    x[scast(usize, ix)],
-                    y[scast(usize, iy)],
+                    x[scast(u32, ix)],
+                    y[scast(u32, iy)],
                     ctx,
                 );
 
@@ -58,13 +58,13 @@ pub fn dot_sub(
         if (comptime types.isArbitraryPrecision(C)) {
             @compileError("zml.linalg.blas.dot_sub not implemented for arbitrary precision types yet");
         } else {
-            for (0..scast(usize, n)) |_| {
+            for (0..scast(u32, n)) |_| {
                 ops.add_( // ret += x[ix] * y[iy]
                     ret,
                     ret.*,
                     ops.mul(
-                        x[scast(usize, ix)],
-                        y[scast(usize, iy)],
+                        x[scast(u32, ix)],
+                        y[scast(u32, iy)],
                         ctx,
                     ) catch unreachable,
                     ctx,
