+++ src/linalg/blas/rot.zig
@@ -7,11 +7,11 @@ const ops = @import("../../ops.zig");
 const blas = @import("../blas.zig");
 
 pub fn rot(
-    n: isize,
+    n: i32,
     x: anytype,
-    incx: isize,
+    incx: i32,
     y: anytype,
-    incy: isize,
+    incy: i32,
     c: anytype,
     s: anytype,
     ctx: anytype,
@@ -26,8 +26,8 @@ pub fn rot(
 
     if (ops.eq(c, 1, .{}) catch unreachable and ops.eq(s, 0, .{}) catch unreachable) return;
 
-    var ix: isize = if (incx < 0) (-n + 1) * incx else 0;
-    var iy: isize = if (incy < 0) (-n + 1) * incy else 0;
+    var ix: i32 = if (incx < 0) (-n + 1) * incx else 0;
+    var iy: i32 = if (incy < 0) (-n + 1) * incy else 0;
     if (comptime types.isArbitraryPrecision(X) or
         types.isArbitraryPrecision(Y) or
         types.isArbitraryPrecision(C) or
@@ -38,17 +38,17 @@ pub fn rot(
         defer ops.deinit(&temp1, ctx);
         var temp2: Ca = try ops.init(Ca, ctx);
         defer ops.deinit(&temp2, ctx);
-        for (0..scast(usize, n)) |_| {
+        for (0..scast(u32, n)) |_| {
             try ops.mul_(
                 &temp1,
                 c,
-                x[scast(usize, ix)],
+                x[scast(u32, ix)],
                 ctx,
             );
             try ops.mul_(
                 &temp2,
                 s,
-                y[scast(usize, iy)],
+                y[scast(u32, iy)],
                 ctx,
             );
             try ops.add_(
@@ -61,12 +61,12 @@ pub fn rot(
             try ops.mul_(
                 &temp2,
                 s,
-                x[scast(usize, ix)],
+                x[scast(u32, ix)],
                 ctx,
             );
 
             try ops.set(
-                &x[scast(usize, ix)],
+                &x[scast(u32, ix)],
                 temp1,
                 ctx,
             );
@@ -74,7 +74,7 @@ pub fn rot(
             try ops.mul_(
                 &temp1,
                 c,
-                y[scast(usize, iy)],
+                y[scast(u32, iy)],
                 ctx,
             );
             try ops.sub_(
@@ -84,7 +84,7 @@ pub fn rot(
                 ctx,
             );
             try ops.set(
-                &y[scast(usize, iy)],
+                &y[scast(u32, iy)],
                 temp1,
                 ctx,
             );
@@ -96,39 +96,39 @@ pub fn rot(
         @compileError("zml.linalg.blas.rot not implemented for arbitrary precision types yet");
     } else {
         var temp: Ca = try ops.init(Ca, .{});
-        for (0..scast(usize, n)) |_| {
+        for (0..scast(u32, n)) |_| {
             ops.add_( // temp = c * x[ix] + s * y[iy]
                 &temp,
                 ops.mul(
                     c,
-                    x[scast(usize, ix)],
+                    x[scast(u32, ix)],
                     ctx,
                 ) catch unreachable,
                 ops.mul(
                     s,
-                    y[scast(usize, iy)],
+                    y[scast(u32, iy)],
                     ctx,
                 ) catch unreachable,
                 ctx,
             ) catch unreachable;
 
             ops.sub_( // y[iy] = c * y[iy] - s * x[ix]
-                &y[scast(usize, iy)],
+                &y[scast(u32, iy)],
                 ops.mul(
                     c,
-                    y[scast(usize, iy)],
+                    y[scast(u32, iy)],
                     ctx,
                 ) catch unreachable,
                 ops.mul(
                     s,
-                    x[scast(usize, ix)],
+                    x[scast(u32, ix)],
                     ctx,
                 ) catch unreachable,
                 ctx,
             ) catch unreachable;
 
             ops.set( // x[ix] = temp
-                &x[scast(usize, ix)],
+                &x[scast(u32, ix)],
                 temp,
                 ctx,
             ) catch unreachable;
